Remove unused code from packaging
In the past, our node_modules depended on a module called `zmq`. Now this module
is named `zeromq` and the executable generated by the electron packager which
uses `zmq.node` no longer depends on `libzmq`.

In the past we had code that tried to find the specific libzmq and for linux
libpgm that was used and copied it inside the generated bundle. Then the wrapper
script also set the `LD_PATH`.

Since the change of requirements this seems to no longer be required.