Compare historical price to USD and adjust to USD if inconsistent
Fix #221. From now on if we get a historical price in a fiat currency
other than USD we always compare to USD. If an incosistency is found
then we take the price adjusted to the USD rate as we consider USD as
the true price.

Also related to #92 since we added a new source of querying historical
forex (fiat to fiat) data. We now query
https://api.exchangeratesapi.io/ for historical forex data and iff
there is a problem with the returned data we then fallback to
cryptocompare. Finally the exchangeratesapi data are cached on disk.