Remove all usages of cost_currency. Now determined by pair format.
With now having only 1 Trade named tuple this commit cleans up some
ambiguity in the code of how pairs describe trades. In the past we had
the cost_currency to describe what is used to buy an asset and what is
received when selling an asset.

Now we make sure that **always** in a BASE_QUOTE pair, we buy the BASE
asset paying with QUOTE and we sell the BASE asset and get QUOTE.

For exchanges where the opposite was happening (Poloniex, Bittrex) the
pair is swapped when the trade is imported.

All tests are also now adjusted.

Also removed the "cost" attribute as it's always amount * rate so it's
redundant to keep this information for each trade.