User settings now keep data cache and properly reloads each visit
Fix #230

Each time the user settings page is loaded, we now render the entire
page again. But we don't query the backend/blockchain each time
now. Instead the data is cached and if it has been queried before it's
used to populate the tables when visiting the user settings again.