Add the first rendition of the known assets mapping
The mapping is added from scanning through kraken's assets. The key is
the canonical symbol and it maps to an object with some data.

The data include a name and a type. Optionally can included an
"active": false entry if the asset is no longer active. They also may
optionally include a "started" to state the timestamp at which the
asset started. If not given it's assumed to be the epoch start.

Also can optionally contain a "forked" entry showing that this asset
forked from another asset.

It can also optionally contain a swapped_for entry showing that this
asset is now inactive but is swapped_for another asset