Last trade is now always taken into account at history creation
Fix #361

The function `limit_trade_list_to_period()` had a bug. It should limit
the sorted list of trades to a time period. But if all trades are
inside the range the last one is skipped. This resulted to always
missing the very last trade when doing the accounting calculation.

This commit fixes the function and unit tests it.