Temporarily skip cryptocompare price history sanity tests
Cryptocompare is changing the way historical data are returned.

Quoting @vcealicu:

We decide on historical trading path based on the current trading volume

If MLN trades a lot of ETH on one day, and you ask for historical MLN-USD, we do it as MLN-ETH × ETH-USD

If next day, it trades more in BTC

We historical as MLN-BTC × BTC-USD.

----

So for this week while this transition is happening results may vary
so the tests should be temporarily skipped until the transition is complete.