Fix bugs in history cache handling
Cache handling in history is now more straightforward but also rather
simple.

If anything is wrong at any point during the reading of any of the
history cache files we bail out, delete all cache files and start
history creation from scratch.

The code is also cleaner. But performance is not the best. See
https://github.com/rotkehlchenio/rotkehlchen/issues/400 for an improvement