Fix the issue where zero-price token removal did not work
Having had a zero price returned for a token (like the PRL in #467)
does not add the token to the totals.

This way when the token removal is attempted, trying to remove it from
the totals results into a key error which interrupts the worker thread
and does not end up removing the token from the DB.

This patch fixes this by simply removing a token from the totals if it
already exists there.

Fix #467