Add ethereum transactions to the DB
Ethereum transactions are now also saved in the database.

Initially for each address in a time range we first check if the last
DB trade is at the limit of the range. If yes then that means the
range has been queried before succesfully and then we do not query
again.

Else we query etherscan and save the transactions in the DB.

The logic needs more finetuning, and needs to be made similar to the
rest of the DB + API querying logic, but unfortunately etherscan
querying uses block numbers and not timestamps. Converting between
timestamps and block numbers is non trivial.