Do not force mutate eth accounts to be checksummed
Fix #519.

This should no longer happen. The very first upgrade of the DB turned
all address entries into checksummed entries. Since then we make sure
to always add addresses as checksummed in the DB and the API also
makes sure that all address input is checksummed.

So this commit closes the chapter by warning the user if a
non-checksummed eth account is found in his DB. But the only way this
happens is if the user manually edits it.