Rest API async query methods now return dicts
Before they were all returning a Tuple[result, message]. And this was
written in the task if if was an async query or returned directly if
not.

Using a dict allows to also return multiple other values from some of
these methods. For example the status code that should be returned
which will be of no use to an async task query but if it's not made
async the information is needed.