Properly disable taxfree_after_period in vue when set in the UI
Can not set a setting to None by providing as `"setting_name": null`
from the API as that is interpreted as skipped setting. Each setting
that can be disabled has its own way and the `taxfree_after_period`
needs to be set to `-1` via the API in order to disable it.