Handle the case where new account with premium uses wrong password
When the user creates a new account and inputs valid premium keys,
then the database will be pulled from the server. In order for that to
work properly the new account should have the same password as the
original password given to the DB in the server.

This commit handles the case where the user gives a different password
more gracefully by providing an error message instead of crashing