Fix flakiness in test_premium.py::test_upload_data_to_server
Fix #856

The test was taking the `now` timestamp after the action where the
time comparison should happen. So it was only passing if the action
took less than a second.

It failed for anything else. The intended usage of the test was to
take the `now` BEFORE the action occured.

Now it should no longer be flaky