Fix flakiness for tests that use a real geth blockchain
The length of the generated test random marker(100 bits) is not guaranteed in
hex. Most of the times it was indeed 24 characters but some times it
was 23. This lead to a failure when comparing with the running marker.

Now we take the random marker's length into account and no longer
compare to a hard coded 24.

Fix #29