Fix flakiness in blockchain tests that remove accounts
At account removals if an asset was expected to be zero the usd value
was also expected to be zero. But due to the way that usd value was
being decreased there could be minor floating point errors.

Right now we changed the implementation so that if asset value is
reduced to zero then usd value is also set to zero automatically.

Fix #879
Fix #871