Search for closest DSR Chi both backwards and forwards
When searching events around a block number for the closest DSR chi
event we search starting from the timestamp's block number and going
both backwards and forwards.

If going forwards we reach the current block, we ask for the current
chi and use that.

If no event is found ~1 months worth of blocks on each side we give
up.