Fix flakiness on test_log_anonymization
Anonymizing logs creates random values for them. And the check was
just checking that the original value is in or not in. But it was
failing with something like amount=55 still being in the logs. But it
was actually a random value starting with 55 like 5564.1 .. or 55.15 e.t.c.

Fix #945