Also catch FileNotFoundError in DB's destructor
Fix #868

This seems to happen only in tests. The destructor for the DB from a
previous test seems to fire up when a new test starts and then it
can't find the DB file of the previous test.

Simply log an error and ignore in that case.

Perhaps we can also somehow force the destructor to fire up at the end
of the test through yielding fixtures but a quick test did not give
any results.