Add in memory-caching for makerdao vault and dsr historical results
Essentially cache the results in memory for 2 hours and return them at
subsequent queries. If any ethereum account is added or removed the
cache is cleaned.

This is not the proper solution and should be combined with saving the
data in the user's database.

DSR caching: https://github.com/rotki/rotki/issues/746
Vault caching: https://github.com/rotki/rotki/issues/1011