Fix long wait time for async aave balance query
The code before was making sure that the ETH balances were queried
before the Aave balance query.

This was not needed. Only the defi balances needed to have been
queried before. And the query was happening before the actual async
task was enqued. Which was a no no.

The code around those queries has been changed to prevent that from
happening.