Attempt to fix negative tax report progress percentage
Fix #1072

Need some help here @kelsos.

Found out that this occurs due to the way the frontend initially sets
the start time of the processing. The start time of processing is not
the same as the start time of the the report. The start time is the
time of the first event processes which is returned by the backend.

This is how the start time of the processsing should be reported to
the frontend.

But I have a problem as it does not seem to actually pick it up with
the changes I made.