Handle the bittrex API v3 pair format
In the previous version of the bittrex API the base currency was the
cost currency and vice versa. So we had to switch the pairs to make
them agree with our trade format.

Since v3 of their API they switched the pairs to the same format we
use so this switch is unnecessary.

Fix #1302