Initial PoC tool for extracting uniswap v2 LPs
As per
https://github.com/rotki/rotki/issues/717#issuecomment-721229173 this
tool uses the method specified both there and in the discord chat of
Rotki.

It's not completely finished.

We allow extraction of LP tokens and their details from both an
ethereum node and the subgraph as a source.

Some of the code needs cleaning up and the encoding/decoding
functionality should be abstracted away.

Also need to play with chunk size for querying of ethereum
node. Etherscan should not be used for a multicall as it has a very
low tolerance due to the url size limit.