Use a hardcoded address for the Zerion DeFi SDK adapter contract
After talking with the Zerion guys I confirmed that if they ever
change the contract the API will also change so by querying
`api.zerion.eth` we run the danger of using a new contract without
having updated our API connector.

So now just use a hardcoded address, and if they ever upgrade and we
want to use the upgraded contract we look at it again then.

Good new is that this simplifies some things in the initialization
process, also for tests