Detect correct validator index of the user under all circumstances
Fix #1887

We were assuming that the deposit index as emitted by the log in the Eth1 chain was also the validator index in the Eth2 chain. It can be, but not necessarily. In fact the validators are put in a queue and are voted in inclusion in a specific order. So the deposit index is not necessarily the validator index. That can only be determined later in the Eth2 chain. (https://kb.beaconcha.in/ethereum-2.0-depositing#the-deposit-contract)