Note when CC rate limits us and avoid it for queries for some time
The main places where cryptocompare is queried is in inquirer for current
prices and price historian for historical prices. And there we have alternatives.

At those places, check if we were recently rate limited by
cryptompare and if we were then just choose the alternative. At the
moment that is coingecko