Handle coingecko rate limiting
This is a kind of hopeful handling. Can still get failed responses due
to rate limiting, but limits them to an absolute minimum.

Need to think if a different approach should happen or if we need to
space the waiting until they reach 60 seconds (which could be kind of
an overkill)