When deriving addresses from xpub use only consecutive indices
Should fix #1759 for the problem that a user reported where inputting
an xpub from scratch gets all addresses but deriving addresses from
xpub over time does not.

The problem seems to be that we derive the initial addresses at Rotki
and then any non-used addresses from the derived ones are never seen
again since the highest derived index is taken.

The solution is to start re-derivation from the highest consecutive
index only, not the highest index in total. This way we should not
miss any addresses.