Also return trade assets in the query_owned_assets
Before we were only returning any assets found in balance
snapshots. Now we also go through all trades and take the assets of
the trade pair as also owned assets.

This also has the nice effect that the cache of cryptocompare queries
is populated with assets that we have had before rotki had the chance
to take balance snapshots