Fix adex test by not returning a usd value + token for withdraw
This is a temporary fix. It seems that the channel id is missing from
the channel_id_token mapping. So it means it does not appear in the
fee rewards mapping

The problem is that the channel withdraw token is now shown as
None from the graph. No idea why. And this causes the logic to be
unable to get a usd value of the token balance withdrawn from the adex channel