Fix botched db upgrade test
We botched the name of the file in the DB upgrade test. The file
nameshould have been forex_history_file.json ->
price_history_forex.json and not the other way around.

The test was passing both locally and in actions only since the cached
directory was used. And this is why it failed on OSX. Due to the lack
of prior caching.

It's a minor thing so just adjusted the test with a note on botching
and removed the cache for the test

Fix #2254