Temporarily change USDT asset type to only ethereum token
Having an asset as "ethereum token and more" is only for semantic
purposes at the moment such as in USDT also being an Omni token.

Changing USDT to be only an ethereum token here so that #2400 can be
fixed for users without actually making a release since their rotki
will automatically pull new all_assets.json and by having USDT as
ethereum only the exception will not be raised.

Once the proper fix for #2400 is in then we should actually revert this