Fix Kucoin trade parsing for some really specific edge cases
In some very specific edge cases Kucoin was returning a string id
which was parseable as a number and we were returning it as an fval
due to rlk_decode_value.

We are now ignoring the relevant kucoin attribute keys and also force
string there so this should not happen again.

Fix #2398