Do not double count balance for Aave v2 tokens not in Aave v1
The usual issue of new things we already detected being also detected
by zerion contract which is not supposed to detect balances.

Fix #2534

The tricky part here is that this only occured for aTokens in v2 but
not in v1. Because those already in v1 were getting stopped by https://github.com/rotki/rotki/issues/2542