Fix all balancer tests. They were all wrong.
It's quite amazing that the rate was wrong in all tests.

All you need to do is read the rate aloud with the rules defined in
our docs and look at the transaction in etherscan to see it's wrong.

Example:

https://etherscan.io/tx/0x1b0d3525964d8e5fbcc0dcdeebcced4bec9017f648e97c3c9761fda1ca6e7b22

Buy 8.0595 WETH for USD at a rate of 1538.7083566 USD per ETH

and not ..

Buy 8.0595 WETH for USD at a rate of 0.00064989573 USD per ETH