Do not query veCRV as a token
Ignore the veCRV balance in token query. It's already detected by
defi SDK as part of locked CRV in Vote Escrowed CRV. Which is the right way
to approach it as there is no way to assign a price to 1 veCRV. It
can be 1 CRV locked for 4 years or 4 CRV locked for 1 year etc.