Fix asset updates from old db withou assets_version key
If the Global DB is upgraded from an old database without an
assets_version key then now the upgrade will always work and will not
be stuck in a loop.

The problem was that if the assets version key was missing we were
not entering it due to using an update statement. Now we use insert or replace