Remove anonymized logs setting and functionality
Fix #2584

This makes no sense as it's not really applicable in a safe and
scalable way for anything that would need anonymizing.

Additionally logs are off by default and as such having anonymization
does not really help anywhere. And the time when you need the logs you
need them to be correct and not anonymized.