Re-invert rate calculation
Getting confused by this. When generating a profit/loos report, the
calculated amounts "Paid in Asset" & "Paid in USD" are incorrect if the
rate is calculated with `base_amount / quote_amount`. The inverse,
`quote_amount / base_amount` yields the correct result for "Paid in
Asset" & "Paid in USD" in the report.