Fix history query for yearn vaults v2
There was an issue where history for addresses will be empty.
The reason is that we avoid querying the subgraph if it was recently queried
and there was a logic to only query information from database for addresses queried
in the subgraph call.

The effect was that if an adress was recently queried it won't retrieve information
from the database and the result would be empty