Adjust globaldb asset updates logic checks
If the remote schema needs a Global DB Schema upgrade to get to, then
quit the assets upgrade process and ask user to upgrade rotki to get
the new assets.

If the remote schema needs a rotki downgrade, to downgrade the global
db upgrade then skip that particular update and let the user know they
need to fix it manually. Perhaps by resetting the global DB assets.