Latest tables **should not** be created before starting DB upgrade
We had a potential bug which could have been the culprit for botched
user DB upgrades of some of our users.

Before all DB upgrading start we were applying the latest sqlite table
creation. Which is very wrong for the case of old upgrades. If the
user is starting a very old upgrade. Say from v2 to v26 then v2 would
start with the DB being populated with all the new tables with foreign
key relations. And it is bound to fail somewhere along the way from 2
to 26. Happened in tests, which is how I noticed it.

Switched to:

- First doing all upgrades (assuming there is a DB - if not a new one
is created in next step)
- Then applying latest tables creation.

Had to adjust tests quite a bit.

And also important. Now each update that relies on tables that were
added during table creation needs to also add said tables during the
upgrade. Thanks to the tests it was easy to spot those and fix them.