Do not add any NFT asset in global DB
We noticed that some users were having foreign key errors due to nfts
being attempted to be added to their global db. This was a logic
error. NFTs should never be attempted to go at the global DB in their
current state. It was happening when adding the user owned assets.

This fixes it and adds a test for it.