Fix premium DB sync size discrepancy in same machine
Fix #2178

The code before was assuming that if the local size is less than the
remote size we should ask the user if they want to pull the remote
data and replace local.

Which is fine in some cases but should not happen if you only use one
system and one rotki instance. Yet it was happening since if you say
remove an ethereum account a lot of data is removed from the DB.

And with less data in the local DB rotki asks if it should pull from
remote. From now on we only deal with timestamp for this prompt of the
DB sync at start.