Binance should die in righteous holy fire
Their API is the shittiest shit out there. Time and time again it has
proven to be the most incompetent, least technical and worst
maintained API of all centralized exchanges.

In the latest installment of their continuous incompetence the fiat
payments endpoint (https://binance-docs.github.io/apidocs/spot/en/#get-fiat-payments-history-user_data)
seems to be a special case in the way it returns data with an empty list.
It may omit the data key completely and return something like this:
{'code': '000000', 'message': 'success', 'success': True, 'total': 0}
which breaks the expectations of our binance api parser and goes
against their own docs.

BURN THAT SHIT WITH FIRE