Change subgraphs for Uniswap and Compound
The uniswap subgraph that we were using was a fork since the official
one was missing a field to correctly filter by the address executing the
transaction.

Now the official subgraph has been fixed and includes this field.
Also we update the compound subgraph address since the issue has been fixed
in the official one.