Rework how trades are initializated from coinbase conversions
This PR modfifies how trades are generated from coinbase transactions. The previous
logic assumed that the fee was paid in the converted asset but as how it was noted
by the user at #3947 if all the owned asset is sold this could lead to incorrect calculations
in the profit and loss report since we were making a wrong calculation of the fee costs.

What this PR modifies is:
1. Now the fee is nominated in the received asset
2. The rate is modified to account for the fee
3. The fee is calculated from the difference in the price of given asset/received asset

Closes #3947