Fixes FTX exchange forgetting subaccount
The reason was that the code for the exchange was expecting the key `ftx_subaccount_name`
but the database was returning this value as `ftx_subaccount` making the initialization
for the subaccount field fail.

The solution applied was to use `ftx_subaccount` everywhere to make the name consistent
around the code.

Closes #3998