Refactor transactions enrichers
The pickle decoder was instead of conventional decoder a enricher for transfers.
For how this part of the code is designed it was not automatically imported but
instead imported as normal function and loaded in the logic of the transaction decoders.

The code worked fine but when packaged there was an error dinamically importing the decoder modules.
This error appears in a combination for how pyinstaller creates the package and how walk_packages
works + the fact that the module was pre loaded at the moment of importing the transaction enricher.

The approach took was to refactor the file structures moving the transfer enrichers to a different folder
and ignoring it when loading the decoder modules.

Additionally as discused in the previous pickle PR the jars contracts will be moved to the globaldb once
the PR with the assets gets merged.