Add default value for name and decimals in spam tokens
For spam tokens we don't waste time trying to query on chain info about the properties of the tokens.
Before the decimals and name were not populated but this is incorrect since leads to a inconsistent state.
This PR sets a default value for those fields that can be later modified