Fix edge case of v31->v32 DB upgrade for kraken users
There is an edge case in the DB upgrade from v31->v32 for a very small
subset of kraken users where the already existing events may contain
sequence indices in such an order that they end up being duplicated
when the duplication protection upgrade code runs.

The solution is simple. You first need to do one iteration of the
entire set of events, to keep a mapping of the all event identifiers
to sequence indices

Fix #4381