Properly handle decoding of special token case information
For some contracts like the WETH contract that are missing name,
symbol and decimals we properly handle the raised exception.

What's more since the decoded token is wrong, we can pre-populate the
token info cache with the special case of the old WETH contract as
many users presumably may hit this, with the proper values.

Fix #4420