Chunk call arguments to multicall
We are using multicall contract in the ENS queries to obtains ENS names for addresses. The problem that
we observed is that when the number of addresses is high (I tested it with 73) the url of the call to
the nodes was so long that it failed to make the request. The solution I propose chunks the query arguments
to the multicall contract to avoid such urls. I've tried different chunk sizes and 20 was the biggest that made
the query work.

Also I've avoided to repeat the same logic for the `multicall_2` contract since we don't use it in a way that
the logic breaks in the same way and could affect the speed of some queries.