Fix the complex segfault case at DB progress callback yielding.
What was happening is the following:

1. Get in the progress callback but before the in_callback lock is acquired.
2. Wait for the lock acquisition and context switch out
.A lot of things happen and the callback gets entered by a lot of other sections
and has the lock acquired. Some of those places do connection.commit
and modify the connection. As per the docs that's a no no.
https://www.sqlite.org/c3ref/progress_handler.html
4. We context switch back to the original callback acquire the lock
and then release it and the callback is exited.
Since the connection has been modified KABOOM SEGFAULT.

The way around this is:

1.To also acquire the callback lock before connection
modifications. So before commit and rollback.
2. If we get in the callback and see that the lock is acquired we do
not wait for it. Instead we exit the callback immediately. This
guarantees that in the case this happens we will not context switch
and exit the callback with the connection having been modified. We
just exit immediately so no modifications can ever happen so long as
we are in the callback.