After merge bugfixes->develop detect_tokens does not need a cursor
The bugfix of one of the crashes was to use different cursor for the
ignored_assets and this makes it a read only operation. So all
dynamics change here.

Also since no place where detect tokens was called from was actually
sharing the cursor, I removed it as an argument completely.