Refactor asset balances and icon endpoints
There is a bug in WSGI pallets/flask#900 that makes the escaped %2F (/) not be properly
interpreted and it becomes a slash. The effect of this bug is that the path for certain values was interpreted as
a different url.

The approach was to use the json body or form to send the identifier avoiding assets identifier in urls.
Also the icon resource was split in two to have the post verb used instead of get to send the asset information.