Modify icon retrieval resource to use get instead of post
This endpoint returns an image and the common way to use this kind of events is using a src tag in
the html that triggers a get in the browser. Using post instead of get is possible bad complicated
as it has been explained by @lukicenturi.

The reason to keep it in two resources is for consistency since the other verbs use json and this
one uses the query args (can't use other thing since is get and there is this bug with the view
args)