Change yearn subgraph to an updated working version
The yearn subgraph stopped indexing and two PRs were made to the repository to fix the issue
but haven't been merged. @rareweasel created this PR https://github.com/yearn/yearn-vaults-v2-subgraph/pull/189
that is deployed at https://thegraph.com/hosted-service/subgraph/rareweasel/yearn-vaults-v2-subgraph-mainnet
and is the one I've put in use in the repo since it seems to be stable.

As discussed the usage of the subgraphs should be removed but until then this will make the functionality keep
working in the app