Do not remove the detected tokens cache after the assets update
I am assuming we had it there in case the assets update changes an
asset and since before it was a list, and now a cache (without foreign
key reference) we would need to purge or update it manually each time
assets change.

In this case, for this upgrade we update this cache during the user DB
upgrade and no identifiers change during global DB upgrade, so
removing it from here is the correct option.

But we would probably need to think of when we would need to purge
this table in the future.