Detect erc721 tokens at tx decoding
This is a stop-gap measure before 1.26.0 to fix up a bit the ERC721
token detection during transaction decoding and to be able to more
properly store them in the DB and decode them for user transactions.

This will most probably need further adjustments to make the code
understand and aware of the different token types and to handle them
appropriately. Both in the DB and in the transaction decoders