Fix database plaintext is locked error
Fix #5038

What was happening is at some very rare case we were getting into the
`maybe_upload_db` through two different greenlets.

Mind you this should not happen easily. Instead of putting a lock as
the task can't be scheduled at the same time and it can only happen if
both a user presses to push to server and the periodic task happen at
the same time we just protect it if it does happen.

We do that by making sure the name is unique each time. Before
`TemporaryDirectory()` was always returning the same directory, we
were trying the same filename for the DB in the directory and then
BOOM.

We fix this by using a uniquely named temporary file each time.