Move BSQ resolution out of inquirer constructor
The problem that was noticed was an interesting one in an old laptop.

The assets in the global DB were quite old. About a year old. When
initializing the global DB upgrade worked fine. But then when
initializing the inquirer BSQ resolution failed since it did not exist
in the DB and as such the entire app crashes and could not open.

That shows that the approach with the resolution is rather brittle and
will need at least in some places rethinking.

But the solution is simple. Move the BSQ resolution only where it's
used. And warn the user very clearly if the 2 base assets USD and WETH
are removed.