Fix potential foreign key error during v34->v35 DB upgrade
Fix #5052

At _update_history_event_assets_identifiers_to_caip_format it was
possible that a history event contained an asset which was missing
from the assets table.

Since the DB upgrade modifies the history_events table to enforce it
so that assets are connected to the assets table this upgrade can
break if for some reason an asset is missing from the asset table.

To fix it we catch the exception and use it to simply add the asset to
the assets table before re-attempting the history event insertion