Fix uploading user DB for windows users with premium subscription
fix #5051

So apparently `tempfile.NamedTemporaryFile` works differently in
Windows than in Unix systems.

Quoting [the
docs](https://docs.python.org/3/library/tempfile.html#tempfile.NamedTemporaryFile):

> Whether the name can be used to open the file a second time, while the
> named temporary file is still open, varies across platforms (it can be
> so used on Unix; it cannot on Windows)

That means that in Windows the temporary file is opened once at the
context manager, stays open and when it's attempted to be opened by
sqlcipher as a DB it goes BOOM BOOM.

The solution for this to work on all platforms is to explicity close
it to allow reopening, to not delete file when closing and to do
cleanup manually at the end.

Kill me.