Introduce cache_response_timewise_immutable
cache_response_timewise saves the cache in a dict, which is
mutable. It returns a reference to it which must not be mutated by the
callers.

We now introduce a cache decorator which creates a copy of the cached
result before returning it.

If any caller wants to mutate the reference to the cache they
should use this decorator.