Disable PEP 517 when packaging the app
After introducing ruff we added a pyproject.toml file and this changed the default
behaviour when packanging the app since the setup.py configuration was being ignored.

The reason to keep that file is that we do an editable install (pip install -e .) that
allows us to call the rotkehlchen module after modifications without needing to install
it again.

This kind of install is allowed with setuptools but from version 64 and we are in 60.7.1
because pyinstaller requires it.

This change allows us to keep the current setup without any additional change and we can
later study how to update setuptools if possible and then move the content of the setup.py
file to the pyproject.toml file