Change the ubuntu version used in CI for tests
Openssl changed the default configuration and the alogorithm used
by some bitcoin libraries was disabled.

Tried to edit the openssl configuration but I couldn't get the CI
working so I opted to downgrade the ubuntu version.

Openssl is studying to reverse this change as it has impacted many
other projects

[run all]