Attempt to fix database plaintext is already in use again
This is a followup from
https://github.com/yabirgb/rotki/commit/a12addbcbd1a8b01019dc9197cda0d7f08389aa8
as the problem seems to keep happening

The idea is that probably the NamedTemporaryFile does not give a
unique name each time and thus the problem persists. So we use ts_now
to force it to always use a unique name.

Also add better logging of the temp db path.