Create per-chain EvmAccountingAggregator and a grouping class
This commit moves the core logic of aggregating per-module evm
accountants into a base class for all evm chains.

Then two different classes are instantiated. One for Ethereum and one
for Optimism. The reason for this is:

1. Each accountant also gets passed the inquirer class of the chain so
it can query things on-chain if it needs to. This would probably make
more sense only if there is archive capability.
2. Subclassing instead of generating two EvmAccountingAggregator
instances with different arguments is due to the potential of having
different logic per evm chain, just like we do with the decoders themselves.