Add decoder for optimisim delegate change
This also touches the "Governance Propose" subtype and generalizes it
to "Governance". It was the only governance related subtype, so this
made sense. The action details are given in the `notes` part of the
event. If we ever want to create per governance actions subtypes we
can revisit this, but it probably won't be necessary.