Make sure curve pools are queried at decoding time
This is a start of generalizing the "Decoder that needs some external
data to exist at decoding time" approach.

- Made a new mixin for Decoders that require this and made sure that
at start of transaction decoding whatever needs to be queried per
decoder is queried.
- Split out the curve protocol cache query from EthereumManager to
EthereumNodeInquirer since this is the only thing that is actually
used to query the chain.
- Made sure that when querying from the decoder the
EthereumNodeinquirer version is used, but if used from outside the
`EthereumManager` is used which also encapsulates an update to the decoder.