Add information about where new tokens are first seen
Add new argument in `get_or_create_evm_token` which allows specifying
where a token newly added to the DB was first seen.

The value can either be missing or it can be a transaction hash or a
string description of an action.

This is used in the websocket messages sent to notify the frontend.