Fresh DB no longer runs all data migrations
In the past we had a new DB run all data migrations. This was not
needed. The only reason it was being done was to do two things:

1. Spam assets addition (Migration: 3)
2. Read and write rpc nodes to the DB (Migrations: 4, 6, 7)

This commit starts a fresh DB at latest migration, so no migrations
run for a fresh DB. Instead those actions are not set apart as actions
to perform to populate a fresh DB.

Tests and their fixtures/mocks are also adjusted properly.