Etherscan ts to blocknumber utilized closest argument properly
This is important as after conversations with etherscan engineers I
realized that the way the API "should" (but not always does) behave
is:

Timestamp BEFORE: should show the first block <= of the provided timestamp (e.g. 1st block of 10 blocks of timestamp 1653220466)
Timestamp AFTER: should be the first block > of the provided timestamp (e.g. 10+1th block of timestamp 1653220466, effectively the first block after timestamp 1653220466)