[UNDO]For all to_block etherscan queries use before instead of after
We started using the "closest" argument of getblocknobytime in
etherscan and had tried to use "before" for from timestamps and
"after" for to timestamps so we get the biggest range possible.

The problem with this approach is that if asking to get block for up to the
current timestamp and giving "after", then etherscan errors out
instead of returning the latest block number. So for now we keep the
functionality as it was before with "before". If they fix it, or if we
change it and add extra checks (probably too complicated?) we can undo
this commit.