Not connected, not syncing exchange query won't query all
Due to a faulty if condition if an exchange that used to be connected
to and is now removed is attempted to be queried, for all trades then
all exchanges would be queried instead if that exchange was also in the
non-syncing exchanges list.

This commit adds a test for and fixes it.